moved approximateWcs to its own file
